refactor: move useUserPosts to entity layer and update related imports
