+++ src/entity/post/model/useSearchPosts.ts
@@ -17,21 +17,3 @@ export const useSearchPosts = () => {
     enabled: !!searchQuery,
   })
 }
-
-//  // 게시물 검색
-//  const searchPosts = async () => {
-//     if (!searchQuery) {
-//       fetchPosts()
-//       return
-//     }
-//     setLoading(true)
-//     try {
-//       const response = await fetch(`/api/posts/search?q=${searchQuery}`)
-//       const data = await response.json()
-//       setPosts(data.posts)
-//       setTotal(data.total)
-//     } catch (error) {
-//       console.error("게시물 검색 오류:", error)
-//     }
-//     setLoading(false)
-//   }
